binutils: Fix x86_64 override usage
This override was applying to binutils-native meaning it would behave differently
depending upon the host it was building on. This is not a good idea and we have
tests to detect it which failed:

oe-selftest -r sstatetests.SStateTests.test_sstate_32_64_same_hash

binutils-native already enables all targets so we can avoid issues by not setting this
option in the native case.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>